Adapt imports and docs to move of base data type classes
