+++ aiida/backends/tests/nodes.py
@@ -1136,7 +1136,8 @@ class TestNodeBasic(AiidaTestCase):
         Test that setting a basetype as an attribute works transparently
         """
         from aiida.orm.data.parameter import ParameterData
-        from aiida.orm.data.base import Str, List
+        from aiida.orm.data.base import List
+        from aiida.orm.data.str import Str
 
         # This one is unstored
         l1 = List()
@@ -1177,7 +1178,8 @@ class TestNodeBasic(AiidaTestCase):
         """
         Test that setting a basetype as an attribute works transparently
         """
-        from aiida.orm.data.base import Str, List
+        from aiida.orm.data.base import List
+        from aiida.orm.data.str import Str
 
         # This one is unstored
         l1 = List()
