+++ aiida/backends/tests/work/job_processes.py
@@ -11,7 +11,7 @@
 from aiida import work
 from aiida.backends.testbase import AiidaTestCase
 from aiida.common.utils import classproperty
-from aiida.orm.data.base import Int
+from aiida.orm.data.int import Int
 from aiida.orm.calculation.job.simpleplugins.templatereplacer import TemplatereplacerCalculation
 from aiida.work.class_loader import ClassLoader
 from aiida.work.job_processes import JobProcess
