+++ aiida/backends/tests/work/test_workfunctions.py
@@ -10,7 +10,8 @@
 import aiida.orm
 import aiida.work.utils as util
 from aiida.backends.testbase import AiidaTestCase
-from aiida.orm.data.base import get_true_node, Int
+from aiida.orm.data.bool import get_true_node
+from aiida.orm.data.int import Int
 from aiida.orm import load_node
 from aiida.work.launch import run, run_get_node
 from aiida.work.workfunctions import workfunction
