+++ docs/source/concepts/expose_examples/run_complex.py
@@ -2,7 +2,9 @@
 
 from __future__ import print_function
 
-from aiida.orm.data.base import Int, Float, Bool
+from aiida.orm.data.bool import Bool
+from aiida.orm.data.float import Float
+from aiida.orm.data.int import Int
 from aiida.work import run
 
 from complex_parent import ComplexParentWorkChain
