+++ docs/source/orm/dev.rst
@@ -140,7 +140,30 @@ XyData
 
 Base data types
 +++++++++++++++
+
+BaseType and NumericType
+------------------------
 .. automodule:: aiida.orm.data.base
+   :members:
+
+Bool
+----
+.. automodule:: aiida.orm.data.bool
+   :members:
+
+Float
+-----
+.. automodule:: aiida.orm.data.float
+   :members:
+
+Int
+---
+.. automodule:: aiida.orm.data.int
+   :members:
+
+Str
+---
+.. automodule:: aiida.orm.data.str
    :members:
